Safari-611.1.10/LayoutTests/canvas/philip/tests/2d.gradient.object.return.html

<!DOCTYPE html>
<title>Canvas test: 2d.gradient.object.return</title>
<script src="../tests.js"></script>
<link rel="stylesheet" href="../tests.css">
<body>
<p id="passtext">Pass</p>
<p id="failtext">Fail</p>
<p class="output">These images should be identical:</p>
<canvas id="c" class="output" width="100" height="50"><p class="fallback">FAIL (fallback content)</p></canvas>

<ul id="d"></ul>
<script>
_addTest(function(canvas, ctx) {

window.CanvasGradient.prototype.thisImplementsCanvasGradient = true;

var g1 = ctx.createLinearGradient(0, 0, 100, 0);
_assertDifferent(g1.addColorStop, undefined, "g1.addColorStop", "undefined");
_assertSame(g1.thisImplementsCanvasGradient, true, "g1.thisImplementsCanvasGradient", "true");

var g2 = ctx.createRadialGradient(0, 0, 10, 0, 0, 20);
_assertDifferent(g2.addColorStop, undefined, "g2.addColorStop", "undefined");
_assertSame(g2.thisImplementsCanvasGradient, true, "g2.thisImplementsCanvasGradient", "true");


});
</script>
